Rebuilt entire repository because security
First couple commits had a .env file that wasn't meant to be public, and the more recent commits had a private key in a file that was not meant to be there. Since there's not really a point in keeping the middle where neither of those were in the project, just rebased the entire project and this is where it's starting. Gotta love git and not being user friendly.